Rename types for consistency
